+++ sample/app/ui-material/material-sample-form.model.ts
@@ -108,6 +108,9 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
         },
         errorMessages: {
             required: "Field is required"
+        },
+        custom: {
+            color: "accent"
         }
     }),