+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 5.2.0
+
+### **Features** 
+
+* `additional` property added to `DynamicFormValueControlModel` for individually configuring UI components 
+(see [**new chapter on form component configuration**](https://github.com/udos86/ng-dynamic-forms/tree/master#form-control-configuration))
+* upgrade to NG Bootstrap `beta.7`
+
+***
+
 # 5.1.2
 
 ### **Features**